inference/requirements.txt
boto3==1.34.39
botocore==1.34.39
jmespath==1.0.1
numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
tzdata==2023.4
urllib3==2.0.7
